Configuring Custom Function SupportResolving Class Conflicts
