+++ src/main/kotlin/org/domaframework/doma/intellij/setting/state/DomaToolsSettings.kt
@@ -13,98 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.domaframework.doma.intellij.common
+package org.domaframework.doma.intellij.setting.state
 
-import com.intellij.openapi.module.Module
-import com.intellij.openapi.roots.ModuleRootManager
-import com.intellij.openapi.vfs.VirtualFile
-import org.jetbrains.jps.model.java.JavaResourceRootType
-import org.jetbrains.jps.model.java.JavaSourceRootType
+import org.domaframework.doma.intellij.setting.SettingComponent
 
-val RESOURCES_META_INF_PATH: String
-    get() = "META-INF"
+interface DomaToolsSettings {
+    fun isModified(component: SettingComponent?): Boolean
 
-class CommonPathParameter(
-    module: Module?,
-) {
-    /**
-     * module base path ex)Absolute path of "/src/main"
-     */
-    var moduleBasePath: VirtualFile? = null
+    fun apply(component: SettingComponent?)
 
-    /**
-     * module source directories ex) Absolute path of "/src/main/java","/src/main/kotlin"
-     */
-    var moduleSourceDirectories: MutableList<VirtualFile> = mutableListOf()
-
-    /**
-     * module resource directory ex)  Absolute path of "/src/main/resources"
-     */
-    var moduleResourceDirectories: MutableList<VirtualFile> = mutableListOf()
-
-    var moduleTestSourceDirectories: MutableList<VirtualFile> = mutableListOf()
-    var moduleTestResourceDirectories: MutableList<VirtualFile> = mutableListOf()
-
-    init {
-        setModuleSourcesFiles(module)
-    }
-
-    private fun setModuleSourcesFiles(module: Module?) {
-        if (module == null) return
-
-        val modulemanager = ModuleRootManager.getInstance(module)
-
-        moduleSourceDirectories.clear()
-        modulemanager?.contentEntries?.firstOrNull()?.let { entry ->
-            moduleBasePath = entry.file
-            entry.sourceFolders.map { folder ->
-                val file = folder.file
-                if (file != null) {
-                    when (folder.rootType) {
-                        JavaSourceRootType.SOURCE -> {
-                            moduleSourceDirectories.add(file)
-                        }
-
-                        JavaSourceRootType.TEST_SOURCE -> {
-                            moduleTestSourceDirectories.add(file)
-                        }
-
-                        JavaResourceRootType.RESOURCE -> {
-                            moduleResourceDirectories.add(file)
-                        }
-
-                        JavaResourceRootType.TEST_RESOURCE -> {
-                            moduleTestResourceDirectories.add(file)
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fun isTest(file: VirtualFile): Boolean {
-        val testSource =
-            moduleTestSourceDirectories.firstOrNull { testSource ->
-                file.path.contains(testSource.path)
-            }
-        if (testSource != null) return true
-
-        return moduleTestResourceDirectories.firstOrNull { testSource ->
-            file.path.contains(testSource.path)
-        } != null
-    }
-
-    fun getResources(file: VirtualFile): MutableList<VirtualFile> =
-        if (isTest(file)) {
-            moduleTestResourceDirectories
-        } else {
-            moduleResourceDirectories
-        }
-
-    fun getSources(file: VirtualFile): MutableList<VirtualFile> =
-        if (isTest(file)) {
-            moduleTestSourceDirectories
-        } else {
-            moduleSourceDirectories
-        }
+    fun reset(component: SettingComponent?)
 }
